Change primary dependency for uc_feeds
The primary dependency for uc_feeds was set to "uc_product". Ubercart was changed so that required dependencies were enforced and hidden, product being one of those dependencies. Instead, the dependency for uc_feeds should be "ubercart" since that is the actual dependency. This is all paraphrased from the original BD UC port maintainer.

Otherwise, uc_feeds throws an error during module enable that uc_feeds needs uc_product to be enabled, even though it technically is already enabled.